delete confirm

+++ examples/pages/template/theme.tpl
@@ -146,7 +146,13 @@ export default {
           this.openCopyForm(item);
           break;
         case 'delete':
-          this.deleteUserThemeByName(item.name);
+          this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+            confirmButtonText: '确定',
+            cancelButtonText: '取消',
+            type: 'warning'
+          }).then(() => {
+            this.deleteUserThemeByName(item.name);
+          }).catch(() => {});
           break;
         default:
           return;
